# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OSdentaire API
  version: 0.3.0
  description: >-
    Modules patients (A101) et agenda / RDV (A102) exposant CRUD sécurisé avec RBAC par entête `x-user-role` et
    canal de polling pour tenir l'UI à jour.
servers:
  - url: https://api.osdentaire.local
security:
  - UserRole: []
components:
  securitySchemes:
    UserRole:
      type: apiKey
      in: header
      name: x-user-role
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error slug (ex. `missing_credentials`).
      required: [error]
    PatientContact:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [phone, email, emergency]
        label:
          type: string
        value:
          type: string
        isPrimary:
          type: boolean
      required: [id, type, label, value, isPrimary]
    ConsentForm:
      type: object
      properties:
        id:
          type: string
          format: uuid
        template:
          type: string
        status:
          type: string
          enum: [pending, signed]
        signedAt:
          type: string
          format: date-time
          nullable: true
        fileUrl:
          type: string
          format: uri
          nullable: true
      required: [id, template, status]
    PatientListItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        reference:
          type: string
        fullName:
          type: string
        status:
          type: string
          enum: [active, archived]
        activeTreatment:
          type: string
          nullable: true
        nextVisit:
          type: string
          format: date-time
          nullable: true
        balance:
          type: number
        primaryContact:
          type: object
          nullable: true
          properties:
            label:
              type: string
            value:
              type: string
        pendingConsents:
          type: integer
      required: [id, reference, fullName, status, balance, pendingConsents]
    PatientDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        reference:
          type: string
        fullName:
          type: string
        preferredName:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        phone:
          type: string
          nullable: true
        activeTreatment:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, archived]
        nextVisit:
          type: string
          format: date-time
          nullable: true
        balance:
          type: number
        notes:
          type: string
          nullable: true
        contacts:
          type: array
          items:
            $ref: '#/components/schemas/PatientContact'
        consentForms:
          type: array
          items:
            $ref: '#/components/schemas/ConsentForm'
      required: [id, reference, fullName, status, balance, contacts, consentForms]
    CreatePatientRequest:
      type: object
      properties:
        fullName:
          type: string
        preferredName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        activeTreatment:
          type: string
        nextVisit:
          type: string
          format: date-time
        notes:
          type: string
        balance:
          type: number
        contacts:
          type: array
          minItems: 1
          items:
            type: object
            properties:
              type:
                type: string
                enum: [phone, email, emergency]
              label:
                type: string
              value:
                type: string
              isPrimary:
                type: boolean
            required: [type, label, value]
        consentForms:
          type: array
          items:
            type: object
            properties:
              template:
                type: string
              version:
                type: string
              status:
                type: string
                enum: [pending, signed]
              signedAt:
                type: string
                format: date-time
            required: [template]
      required: [fullName, phone, contacts]
    CreateAppointmentRequest:
      type: object
      properties:
        providerId:
          type: string
          format: uuid
        roomId:
          type: string
          format: uuid
        patientId:
          type: string
          format: uuid
        title:
          type: string
        startAt:
          type: string
          format: date-time
        durationMinutes:
          type: integer
          minimum: 5
          maximum: 240
        notes:
          type: array
          items:
            type: object
            properties:
              body:
                type: string
            required: [body]
      required: [providerId, roomId, patientId, title, startAt]
    CancelAppointmentRequest:
      type: object
      properties:
        reason:
          type: string
      additionalProperties: false
    AgendaResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Appointment'
        providers:
          type: array
          items:
            $ref: '#/components/schemas/ProviderSummary'
        rooms:
          type: array
          items:
            $ref: '#/components/schemas/RoomSummary'
        meta:
          type: object
          properties:
            start:
              type: string
              format: date-time
            end:
              type: string
              format: date-time
            timezone:
              type: string
            cursor:
              type: integer
          required: [start, end, timezone, cursor]
      required: [data, providers, rooms, meta]
    ProviderSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
        initials:
          type: string
        specialty:
          type: string
          nullable: true
        role:
          type: string
        color:
          type: string
        nextAvailableAt:
          type: string
          format: date-time
          nullable: true
      required: [id, fullName, initials, role, color]
    RoomSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        color:
          type: string
        floor:
          type: string
          nullable: true
        equipment:
          type: string
          nullable: true
      required: [id, name]
    AppointmentNote:
      type: object
      properties:
        id:
          type: string
          format: uuid
        authorRole:
          type: string
        authorName:
          type: string
        kind:
          type: string
          enum: [note, notification]
        body:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, authorRole, authorName, kind, body, createdAt]
    Appointment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        status:
          type: string
          enum: [scheduled, cancelled]
        timezone:
          type: string
          example: Europe/Paris
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        slotMinutes:
          type: integer
        provider:
          type: object
          properties:
            id:
              type: string
              format: uuid
            fullName:
              type: string
            initials:
              type: string
            color:
              type: string
          required: [id, fullName, initials, color]
        room:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            color:
              type: string
          required: [id, name, color]
        patient:
          type: object
          properties:
            id:
              type: string
              format: uuid
            fullName:
              type: string
            reference:
              type: string
          required: [id, fullName, reference]
        cancelReason:
          type: string
          nullable: true
        canceledAt:
          type: string
          format: date-time
          nullable: true
        createdBy:
          type: string
        createdByRole:
          type: string
        notes:
          type: array
          items:
            $ref: '#/components/schemas/AppointmentNote'
      required: [id, title, status, timezone, startAt, endAt, provider, room, patient, createdBy, createdByRole, notes]
    AppointmentEvent:
      type: object
      properties:
        type:
          type: string
          enum: [appointment.created, appointment.cancelled]
        data:
          $ref: '#/components/schemas/Appointment'
        cursor:
          type: integer
      required: [type, data, cursor]
paths:
  /api/patients:
    get:
      summary: Liste paginée des patients
      parameters:
        - in: query
          name: q
          schema:
            type: string
          description: Filtre sur nom, référence ou téléphone.
        - in: query
          name: status
          schema:
            type: string
            enum: [active, archived]
            default: active
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Liste paginée
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PatientListItem'
                  meta:
                    type: object
                    properties:
                      total:
                        type: integer
                      limit:
                        type: integer
                      offset:
                        type: integer
                required: [data, meta]
        '401':
          description: Auth manquante
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: RBAC interdit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Paramètres invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erreur interne
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Crée un patient + consentements
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePatientRequest'
      responses:
        '201':
          description: Patient créé
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/PatientDetail'
        '401': { description: Auth manquante, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '403': { description: RBAC interdit, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '422': { description: Charge utile invalide, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '500': { description: Erreur interne, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /api/patients/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Récupère un patient avec contacts/consentements
      responses:
        '200':
          description: Patient détaillé
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/PatientDetail'
        '401': { description: Auth manquante, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '403': { description: RBAC interdit, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: Introuvable, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '422': { description: Paramètre invalide, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '500': { description: Erreur interne, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /api/appointments:
    get:
      summary: Liste les rendez-vous (agenda hebdo, fuseau Europe/Paris)
      parameters:
        - in: query
          name: start
          schema:
            type: string
            format: date-time
          description: Début de plage (UTC). Défaut = lundi courant 00:00 Europe/Paris.
        - in: query
          name: end
          schema:
            type: string
            format: date-time
          description: Fin de plage (UTC). Défaut = start + 7 jours.
        - in: query
          name: providerId
          schema:
            type: string
          description: Liste d'UUID séparés par virgule pour filtrer les praticiens.
        - in: query
          name: roomId
          schema:
            type: string
          description: UUID(s) de salle séparés par virgule.
        - in: query
          name: includeNotes
          schema:
            type: boolean
          description: Inclure les notes/notifications associées.
      responses:
        '200':
          description: Agenda hebdomadaire
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgendaResponse'
        '401': { description: Auth manquante, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '403': { description: RBAC interdit, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '422': { description: Paramètres invalides, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '500': { description: Erreur interne, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    post:
      summary: Crée un rendez-vous et déclenche une notification interne
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppointmentRequest'
      responses:
        '201':
          description: Rendez-vous créé
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Appointment'
        '401': { description: Auth manquante, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '403': { description: RBAC interdit, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '409': { description: Double-booking détecté, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '422': { description: Charge utile invalide, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '500': { description: Erreur interne, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /api/appointments/{id}/cancel:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
    patch:
      summary: Annule un rendez-vous (RBAC praticien/admin)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelAppointmentRequest'
      responses:
        '200':
          description: Rendez-vous annulé
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Appointment'
        '401': { description: Auth manquante, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '403': { description: RBAC interdit, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: RDV introuvable, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '422': { description: Paramètres invalides, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '500': { description: Erreur interne, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /api/appointments/updates:
    get:
      summary: Canal de polling long (30s) pour suivre les créations/annulations
      parameters:
        - in: query
          name: cursor
          schema:
            type: integer
          description: Curseur retourné par le dernier appel (0 pour initialiser).
      responses:
        '200':
          description: Liste d'événements depuis le curseur
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/AppointmentEvent'
                  cursor:
                    type: integer
                required: [events, cursor]
        '401': { description: Auth manquante, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '403': { description: RBAC interdit, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    delete:
      summary: Soft delete (droit à l'oubli)
      responses:
        '204': { description: Archivé }
        '401': { description: Auth manquante, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '403': { description: Admin requis, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: Introuvable, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '422': { description: Paramètre invalide, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '500': { description: Erreur interne, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }